refactor(leet): split the inspector into input and view files
Splits inspector.go (940 lines) along its section markers:
inspectorhandlers.go takes keyboard and mouse input plus pane resizing,
inspectorview.go takes rendering. The model, its lifecycle, and the data
flow stay in inspector.go.
